backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Eric Zhang <zhangeric991@gmail.com>"]
readme = "README.md"
packages = [
    { include = "interview_gpt" },
    { include = "interview_gpt/**/*.py" },
]


[tool.poetry.dependencies]
python = "<3.13, >= 3.9"
cuid2 = "^2.0.0"
pyaudio = "^0.2.14"
scipy = "^1.11.3"
requests = "^2.31.0"
torch = "^2.1.0"
numpy = "^1.26.1"
sounddevice = "^0.4.6"
soundcard = "^0.4.2"
openai = "^1.1.1"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.4"
pypdf = "^3.17.0"
pdfminer-six = "^20221105"
soundfile = "^0.12.1"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
tiktoken = "^0.5.1"
typer = "^0.9.0"
termcolor = "^2.3.0"
pytest = "^7.4.3"
mypy = "^1.6.1"
ruff = "^0.1.4"
tenacity = "^8.2.3"
loguru = "^0.7.2"
fastapi = {extras = ["all"], version = "^0.104.1"}
websockets = "^12.0"
pydantic-settings = "^2.1.0"
python-socketio = "^5.10.0"
webrtcvad = "^2.0.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
